fix(mobile): render workspace images in file markdown previews
