+++ apps/mobile/src/features/files/ThreadFilesRouteScreen.tsx
@@ -1,6 +1,7 @@
 import { NativeHeaderToolbar, NativeStackScreenOptions } from "../../native/StackHeader";
 import { StackActions, useNavigation, type StaticScreenProps } from "@react-navigation/native";
-import { useCallback, useEffect, useRef, useState } from "react";
+import { classifyMarkdownImageSource } from "@t3tools/client-runtime/markdown-images";
+import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { ActivityIndicator, Platform, View } from "react-native";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
 import Svg, { Defs, LinearGradient, Rect, Stop } from "react-native-svg";
@@ -16,10 +17,15 @@ import { SymbolView } from "../../components/AppSymbol";
 import { AppText as Text, AppTextInput as TextInput } from "../../components/AppText";
 import { EmptyState } from "../../components/EmptyState";
 import { LoadingScreen } from "../../components/LoadingScreen";
+import {
+  ThreadMarkdownImage,
+  ThreadMarkdownImageUnavailable,
+} from "../../components/markdownImages";
 import { resolveFileSelectionNavigationAction } from "../../lib/adaptive-navigation";
 import { copyTextWithHaptic } from "../../lib/copyTextWithHaptic";
 import { tryOpenExternalUrl } from "../../lib/openExternalUrl";
 import { useThemeColor } from "../../lib/useThemeColor";
+import type { MarkdownImageRenderer } from "../../native/SelectableMarkdownText";
 import { useThreadSelection } from "../../state/use-thread-selection";
 import { useSelectedThreadWorktree } from "../../state/use-selected-thread-worktree";
 import { useEnvironmentQuery } from "../../state/query";
@@ -50,6 +56,7 @@ import {
   isImagePreviewFile,
   isMarkdownPreviewFile,
   isSvgImagePreviewFile,
+  resolveWorkspaceFilePath,
 } from "./filePath";
 import { useWorkspaceFileAssetUrl } from "./workspaceFileAssetUrl";
 
@@ -94,6 +101,7 @@ function FileContent(props: {
   readonly initialLine: number | null;
   readonly truncated: boolean;
   readonly onRefresh?: () => Promise<void> | void;
+  readonly renderImage?: MarkdownImageRenderer;
 }) {
   const isMarkdown = isMarkdownPreviewFile(props.relativePath);
   const isBrowserFile = isBrowserPreviewFile(props.relativePath);
@@ -145,7 +153,11 @@ function FileContent(props: {
         </View>
       ) : null}
       {props.activeMode === "preview" && isMarkdown ? (
-        <FileMarkdownPreview markdown={props.fileContents} onRefresh={props.onRefresh} />
+        <FileMarkdownPreview
+          markdown={props.fileContents}
+          onRefresh={props.onRefresh}
+          renderImage={props.renderImage}
+        />
       ) : (
         <SourceFileSurface
           contents={props.fileContents}
@@ -511,6 +523,35 @@ export function ThreadFileScreen(props: ThreadFileRouteScreenProps) {
       : null,
   );
   const fileData = fileQuery.data as ProjectReadFileResult | null;
+  const renderMarkdownImage = useMemo<MarkdownImageRenderer | undefined>(() => {
+    if (environmentId === null || threadId === null || relativePath === null || cwd === null) {
+      return undefined;
+    }
+    const lastSeparator = Math.max(relativePath.lastIndexOf("/"), relativePath.lastIndexOf("\\"));
+    // Relative image sources in a rendered markdown file resolve against its own directory.
+    const imageBaseDir =
+      lastSeparator >= 0
+        ? resolveWorkspaceFilePath(cwd, relativePath.slice(0, lastSeparator))
+        : cwd;
+
+    return (image) => {
+      const imageSource = classifyMarkdownImageSource(image.href, imageBaseDir);
+      if (imageSource._tag === "Direct") {
+        return null;
+      }
+      if (imageSource._tag === "Blocked") {
+        return <ThreadMarkdownImageUnavailable alt={image.alt} />;
+      }
+      return (
+        <ThreadMarkdownImage
+          environmentId={environmentId}
+          threadId={threadId}
+          path={imageSource.path}
+          alt={image.alt}
+        />
+      );
+    };
+  }, [cwd, environmentId, relativePath, threadId]);
 
   const handleSelectFile = useCallback(
     (path: string) => {
@@ -665,6 +706,7 @@ export function ThreadFileScreen(props: ThreadFileRouteScreenProps) {
           fileError={fileQuery.error}
           initialLine={targetLine}
           relativePath={relativePath}
+          renderImage={renderMarkdownImage}
           truncated={fileData?.truncated ?? false}
           onRefresh={() => fileQuery.refresh()}
         />
